requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
strawberry-graphql>=0.200.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pytest>=7.0.0
websockets>=11.0.3
requests>=2.31.0
httpx>=0.24.0
